griptape_chat_memory_agent/requirements.txt
python-dotenv==1.0.1
griptape==1.3.0
